Revamp build
