+++ .gitignore
@@ -1,39 +1,10 @@
-.*
-
-!.gitignore
-!.scalafmt.conf
-bin
 .bloop/
-.cache
-/.classpath
-db
-dist
-.DS_Store
-**/*.DS_Store
-eclipse
-*.eml
-.history
-.idea
-.idea_modules
-/.idea_modules
-*.iml
-lib
-lib_managed
-log
-logs
+.bsp/
+.idea/
+.idea_modules/
 .metals/
-modules
-/out
-precompiled
-/.project
-project/metals.sbt
-project/project
-project/target
-/RUNNING_PID
-server.pid
-/.settings
-target
-.target
-test-result
-tmp
 .vscode/
+logs/
+target/
+metals.sbt
+*.class
