+++ .scalafmt.conf
@@ -1,0 +1,11 @@
+version = 2.7.5
+
+preset = defaultWithAlign
+
+maxColumn = 100
+
+continuationIndent = {
+    defnSite = 2
+    callSite = 2
+    ctorSite = 2
+}
