--- app/uk/gov/hmrc/commontransitconventionguaranteebalance/controllers/MicroserviceHelloWorldController.scala
@@ -1,15 +1,0 @@
-package uk.gov.hmrc.commontransitconventionguaranteebalance.controllers
-
-import uk.gov.hmrc.play.bootstrap.backend.controller.BackendController
-import play.api.mvc.{Action, AnyContent, ControllerComponents}
-import javax.inject.{Inject, Singleton}
-import scala.concurrent.Future
-
-@Singleton()
-class MicroserviceHelloWorldController @Inject()(cc: ControllerComponents)
-    extends BackendController(cc) {
-
-  def hello(): Action[AnyContent] = Action.async { implicit request =>
-    Future.successful(Ok("Hello world"))
-  }
-}
