+++ build.sbt
@@ -1,26 +1,48 @@
-import uk.gov.hmrc.DefaultBuildSettings.integrationTestSettings
-import uk.gov.hmrc.sbtdistributables.SbtDistributablesPlugin.publishingSettings
+import scoverage.ScoverageKeys
+import uk.gov.hmrc.DefaultBuildSettings
 
 val appName = "common-transit-convention-guarantee-balance"
 
-val silencerVersion = "1.7.3"
-
 lazy val microservice = Project(appName, file("."))
-  .enablePlugins(play.sbt.PlayScala, SbtDistributablesPlugin)
-  .settings(
-    majorVersion                     := 0,
-    scalaVersion                     := "2.12.13",
-    libraryDependencies              ++= AppDependencies.compile ++ AppDependencies.test,
-    // ***************
-    // Use the silencer plugin to suppress warnings
-    scalacOptions += "-P:silencer:pathFilters=routes",
-    libraryDependencies ++= Seq(
-      compilerPlugin("com.github.ghik" % "silencer-plugin" % silencerVersion cross CrossVersion.full),
-      "com.github.ghik" % "silencer-lib" % silencerVersion % Provided cross CrossVersion.full
-    )
-    // ***************
-  )
-  .settings(publishingSettings: _*)
   .configs(IntegrationTest)
-  .settings(integrationTestSettings(): _*)
-  .settings(resolvers += Resolver.jcenterRepo)
+  .enablePlugins(play.sbt.PlayScala, SbtDistributablesPlugin)
+  .settings(SbtDistributablesPlugin.publishingSettings)
+  .settings(DefaultBuildSettings.integrationTestSettings())
+  .settings(scalacSettings)
+  .settings(scoverageSettings)
+  .settings(
+    majorVersion := 0,
+    scalaVersion := "2.12.13",
+    resolvers += Resolver.jcenterRepo,
+    libraryDependencies ++= AppDependencies.compile ++ AppDependencies.test
+  )
+
+lazy val scalacSettings = Def.settings(
+  // Disable warnings arising from generated routing code
+  scalacOptions += "-Wconf:src=routes/.*:silent",
+  // Disable fatal warnings and warnings from discarding values
+  scalacOptions ~= { opts =>
+    opts.filterNot(Set("-Xfatal-warnings", "-Ywarn-value-discard"))
+  }
+)
+
+lazy val scoverageSettings = Def.settings(
+  Test / parallelExecution := false,
+  ScoverageKeys.coverageMinimumStmtTotal := 90,
+  ScoverageKeys.coverageFailOnMinimum := true,
+  ScoverageKeys.coverageHighlighting := true,
+  ScoverageKeys.coverageExcludedPackages := Seq(
+    "<empty>",
+    "Reverse.*",
+    ".*(config|views.*)",
+    ".*(BuildInfo|Routes).*"
+  ).mkString(";"),
+  ScoverageKeys.coverageExcludedFiles := Seq(
+    "<empty>",
+    "Reverse.*",
+    ".*BuildInfo.*",
+    ".*javascript.*",
+    ".*Routes.*",
+    ".*GuiceInjector"
+  ).mkString(";")
+)
