+++ run_all_tests.sh
@@ -1,0 +1,2 @@
+#!/usr/bin/env bash
+sbt clean coverage test it:test coverageOff coverageReport
